feat: extract AgentController from index.ts
Moves the query loop and all REPL controller logic out of index.ts into
a new AgentController class in src/agent/controllers/agent-controller.ts.
index.ts is now a ~30-line composition root that constructs objects and
calls agentController.start().

What moved into AgentController:
- runQuery() — Claude SDK invocation, session ID management, tool
  permission callbacks (handleAskUserQuestion, handleToolPermission),
  status bar lifecycle, ^C interrupt handling
- start() — startup banner, stdin setup, workspace construction,
  command registration (exit, clear, model, effort, workspace:*, worker:*),
  and the main prompt loop
- logFull() — private module-level helper for repl.log

Tests updated to import AgentController directly and use
vi.spyOn(controller, 'runQuery') instead of injecting a runQueryFn.

Closes #710
