+++ CLAUDE.md
@@ -17,15 +17,15 @@ MVC structure. Models own state; controllers handle external inputs.
 
 ### Agent/Worker (`src/agent/`)
 
-A unified REPL + worker loop. `main()` in `index.ts` runs both interactive and worker modes; `startWorkerMode(display)` in `controllers/worker-controller.ts` sets up the WebSocket session and returns before `main()` takes over the query loop.
+A unified REPL + worker loop. `AgentController` in `controllers/agent-controller.ts` owns the query loop; `index.ts` is a thin composition root (~30 lines) that constructs objects and calls `agentController.start()`.
 
 Follows MVC with three subdirectories:
 
 - **Models** (`models/`) — `workspace.ts` (git/npm workspace management; extends `EventEmitter`, emits `"clone-start"`, `"reset-start"`, `"reset-retry"`, `"reset-reclone"`, `"npm-install"`, `"destroy"`, `"prune-start"`, `"prune-remove"`, `"prune-skip"` events instead of calling `print()` directly — callers subscribe in `startWorkerMode()`; `prune()` is an instance method, not static), `settings.ts` (pure model for runtime-settable preferences; `Settings` is the only value export — owns model and effort state, emits `"change"` on updates via `_setModel`/`_setEffort`; effort/model constants and cache helpers (`EFFORT_LEVELS`, `VALID_EFFORT_VALUES`, `getCachedModels`, `setCachedModels`, `_resetCachedModels`, `findModel`) are all static members of `Settings` — no interactive methods), `query-stats.ts` (tracks token usage and turn counts; exports `QueryStats` class only — `WORKING_VERBS` and `pickWorkingVerb` are private to the module; no view imports)
 - **Views** (`views/`) — `display.ts` (TUI rendering; exports `Display` class with `BrunelConfig` and `StatusBar` injected in constructor — class methods use `this.config.verbose` and `this.statusBar`; `display.clearBreak()` and `display.fmtArgs(input)` delegate to module-level utilities; all pure utilities like `c`, `s`, `resolve`, `fmtEvent`, etc.; no singletons or standalone delegates), `status-bar.ts` (status bar state and rendering; exports `StatusBar` class only — holds all worker status state; constructed in `index.ts` startup and injected into `Display` and `startWorkerMode`; no singleton or `initStatusBar` export), `input.ts` (readline-based REPL input: `ask`, `pick`, `pickMultiple`, `pickQuestion`)
-- **Controllers** (`controllers/`) — `worker-controller.ts` (WS protocol + task lifecycle; defines `WorkerDisplay` interface `{ print, printForemanMessage }` — used by all controllers; owns `WorkerSession` with `session.confirmTaskQuit(info, pickFn?)` as a method, `generateAgentId`, `WORKER_NAMES`, `startWorkerMode(display, statusBar)`, `registerWorkerCommands(session, registry, display)`), `workspace-controller.ts` (registers `/workspace:*` slash commands; `registerWorkspaceCommands(workspace, registry, display)` takes injected `display: WorkerDisplay`), `command-controller.ts` (`CommandRegistry` class — registration only (`register`, `scoped`, `lookup`, `listAll`, `execute`, `_reset`); `CommandController` class — has-a `CommandRegistry` (not extends), adds `dispatch`, `parseSlashCommand`, `listCommandNames`, `listCommands`, `suggest`; exposes `controller.registry` getter for registration in tests; `index.ts` creates `new CommandRegistry()` for registration and `new CommandController(registry)` for dispatch; pass `CommandRegistry` to functions that only register commands), `settings-controller.ts` (`SettingsController` class — wraps `Settings` model with `display: WorkerDisplay`; exposes `pickModel(args, pickFn, fetchModelsFn)` and `pickEffort(args, pickFn)`)
+- **Controllers** (`controllers/`) — `agent-controller.ts` (`AgentController` class — owns the Claude SDK query loop; constructor takes `(display, permConfig, settings)`; `runQuery(prompt, sessionId, ac?, model?, effort?)` runs a single turn via the SDK (handles `AskUserQuestion`/tool-permission callbacks, session ID capture, status bar lifecycle, ^C interrupt); `start(runWorkerMode?, workspaceCfg?)` sets up worker mode if needed, prints the startup banner, registers all REPL commands (`exit`, `clear`, `model`, `effort` plus workspace and worker sub-commands), and drives the prompt loop until exit; `logFull` is a private module-level helper that appends to `repl.log`), `worker-controller.ts` (WS protocol + task lifecycle; defines `WorkerDisplay` interface `{ print, printForemanMessage }` — used by all controllers; owns `WorkerSession` with `session.confirmTaskQuit(info, pickFn?)` as a method, `generateAgentId`, `WORKER_NAMES`, `startWorkerMode(display, statusBar)`, `registerWorkerCommands(session, registry, display)`), `workspace-controller.ts` (registers `/workspace:*` slash commands; `registerWorkspaceCommands(workspace, registry, display)` takes injected `display: WorkerDisplay`), `command-controller.ts` (`CommandRegistry` class — registration only (`register`, `scoped`, `lookup`, `listAll`, `execute`, `_reset`); `CommandController` class — has-a `CommandRegistry` (not extends), adds `dispatch`, `parseSlashCommand`, `listCommandNames`, `listCommands`, `suggest`; exposes `controller.registry` getter for registration in tests; `index.ts` creates `new CommandRegistry()` for registration and `new CommandController(registry)` for dispatch; pass `CommandRegistry` to functions that only register commands), `settings-controller.ts` (`SettingsController` class — wraps `Settings` model with `display: WorkerDisplay`; exposes `pickModel(args, pickFn, fetchModelsFn)` and `pickEffort(args, pickFn)`)
 
-Other key files: `index.ts` (entrypoint; constructs `settings`, `statusBar`, and `display` in order at startup — `const statusBar = new StatusBar({ agentId, settings })`, `const display = new Display(config, statusBar)` — and passes them to all controllers; `runQuery(display, permConfig, ...)` and `main(runQueryFn, permConfig, display, settings, ...)` receive display and settings as parameters), `worker-prompts.ts` (prompt templates).
+Other key files: `index.ts` (thin composition root; constructs `settings`, `statusBar`, `display`, and `AgentController` — then calls `agentController.start(runWorkerMode, workspaceCfg)`; ~30 lines), `worker-prompts.ts` (prompt templates).
 
 ### Shared
